[CRT-382] Reuse error code instead of custom error
