+++ frontend/src/api/fetch.ts
@@ -1,35 +1,5 @@
+import { ApiError, ConflictError } from "@/errors/api";
 import { authenticationStateKey, backendHostName } from "@/utilities/constants";
-
-export class ApiError extends Error {
-  public readonly errorCode?: string;
-
-  constructor(
-    public readonly status: number,
-    message: string,
-    errorCode?: string,
-  ) {
-    super(message);
-    this.name = "ApiError";
-    this.errorCode = errorCode;
-  }
-}
-
-export class ConflictError extends ApiError {
-  constructor(message: string = "Conflict", errorCode?: string) {
-    super(409, message, errorCode);
-    this.name = "ConflictError";
-  }
-}
-
-export class DuplicateReferenceSolutionError extends ApiError {
-  constructor(
-    message: string = "Duplicate reference solution",
-    errorCode?: string,
-  ) {
-    super(400, message, errorCode);
-    this.name = "DuplicateReferenceSolutionError";
-  }
-}
 
 export const fetchApi = async <T>(
   url: string,
@@ -62,15 +32,11 @@ export const fetchApi = async <T>(
 
   if (response.status >= 400) {
     const errorBody = await response.json();
-    if (errorBody?.errorCode === "DUPLICATE_REFERENCE_SOLUTION") {
-      throw new DuplicateReferenceSolutionError(
-        errorBody?.message,
-        errorBody?.errorCode,
-      );
-    }
+
     throw new ApiError(
       response.status,
-      `Unexpected response code ${response.status}`,
+      errorBody?.message,
+      errorBody?.errorCode,
     );
   }
