+++ frontend/src/components/task/TaskActions.tsx
@@ -3,10 +3,10 @@ import { LuTrash } from "react-icons/lu";
 import { useState } from "react";
 import { useRouter } from "next/router";
 import { useDeleteTask } from "@/api/collimator/hooks/tasks/useDeleteTask";
-import { ConflictError } from "@/api/fetch";
 import { getErrorMessageDescriptor } from "@/errors/errorMessages";
 import { ButtonMessages } from "@/i18n/button-messages";
 import { useTask } from "@/api/collimator/hooks/tasks/useTask";
+import { ConflictError } from "@/errors/api";
 import DropdownMenu from "../DropdownMenu";
 import { toaster } from "../Toaster";
 import { Modal } from "../form/Modal";
