+++ frontend/src/components/task/TaskTable.tsx
@@ -13,9 +13,9 @@ import { useDeleteTask } from "@/api/collimator/hooks/tasks/useDeleteTask";
 import { ExistingTask } from "@/api/collimator/models/tasks/existing-task";
 import { capitalizeString } from "@/utilities/strings";
 import { isClickOnRow } from "@/utilities/table";
-import { ConflictError } from "@/api/fetch";
 import { getErrorMessageDescriptor } from "@/errors/errorMessages";
 import { AuthenticationContext } from "@/contexts/AuthenticationContext";
+import { ConflictError } from "@/errors/api";
 import SwrContent from "../SwrContent";
 import ConfirmationModal from "../modals/ConfirmationModal";
 import { ChakraDataTable, ColumnSize } from "../ChakraDataTable";
